% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typesPathways.R
\name{addContraSide}
\alias{addContraSide}
\title{Functions to deal with crossing to the bad side of the brain}
\usage{
addContraSide(connTable)
}
\description{
Functions to deal with crossing to the bad side of the brain
}
