% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivityCluster.R
\name{clusterBag}
\alias{clusterBag}
\title{Cluster neurons in a bag}
\usage{
clusterBag(
  nBag,
  ROIs = NULL,
  clusterOn = c("inputs", "outputs", "both"),
  unit = c("type", "neuron", "supertype1", "supertype2", "supertype3", "databaseType"),
  distance = cos_dist,
  knownStats = FALSE
)
}
\arguments{
\item{nBag}{a neuron bag}

\item{ROIs}{Which ROIs to consider. If NULL (the default) uses all the ROI present}

\item{clusterOn}{whether to cluster on the input vector, the output vector, or both}

\item{unit}{Unit on which the clustering is to be made (neuron or type, or various supertypes)}

\item{distance}{Which distance measure to use (defaults to \code{\link{cos_dist}})}

\item{knownStats}{Whether or not to use \code{knownOutputContribution} and \code{knownWeightRelative} instead of the standard measures}
}
\value{
an object of class \code{connectivityCluster}. The object is a list with components:
\describe{
  \item{inputsTable}{The table of input connections to the neurons being clustered (can be NULL if clustered on outputs)}
  \item{outputsTable}{The table of output connections to the neurons being clustered (can be NULL if clustered on inputs)}
  \item{distance}{object containing a dissimilarity measure between the neurons being considered}
  \item{hc}{\code{\link{hclust}} object, result of clustering}
  \item{unit}{Unit on which the clustering was made (neuron or type, or various supertypes)}
}
}
\description{
Cluster neurons in a bag
}
\seealso{
the lower level \code{\link{connectivityCluster}}
}
