% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rois.R
\name{combineRois}
\alias{combineRois}
\title{Sum a set of ROIs in a connection table or neuronBag}
\usage{
combineRois(connections, rois, newRoi, ...)
}
\arguments{
\item{connections}{A raw connection table or a neuronBag}

\item{rois}{The ROIs to combined}

\item{newRoi}{The name of the newly formed ROI}

\item{...}{Parameters to be passed to getTypeToTypeTable}
}
\description{
Sum a set of ROIs in a connection table or neuronBag
}
