% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivityMatrices.R
\name{conn_mat2df}
\alias{conn_mat2df}
\title{Turn a connectivity matrix back into a data.frame}
\usage{
conn_mat2df(connMat, orderIn = NULL, orderOut = NULL, thresh = 0)
}
\description{
Turn a connectivity matrix back into a data.frame
}
