% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivityCluster.R
\name{connectivityCluster}
\alias{connectivityCluster}
\alias{clusterBag}
\title{Clustering neurons based on connectivity tables}
\usage{
connectivityCluster(
  inputsTable = NULL,
  outputsTable = NULL,
  ROIs = NULL,
  unit = c("type", "neuron", "supertype1", "supertype2", "supertype3", "databaseType"),
  distance = cos_dist,
  knownStats = FALSE
)

clusterBag(
  nBag,
  ROIs = NULL,
  clusterOn = c("inputs", "outputs", "both"),
  unit = c("type", "neuron", "supertype1", "supertype2", "supertype3", "databaseType"),
  distance = cos_dist,
  knownStats = FALSE
)
}
\arguments{
\item{inputsTable}{The table of input connections to the neurons to cluster}

\item{outputsTable}{The table of input connections to the neurons to cluster}

\item{ROIs}{Which ROIs to consider. If NULL (the default) uses all the ROI present}

\item{unit}{Unit on which the clustering is to be made (neuron or type, or various supertypes)}

\item{distance}{Which distance measure to use (defaults to \code{\link{cos_dist}})}

\item{knownStats}{Whether or not to use \code{knownOutputContribution} and \code{knownWeightRelative} instead of the standard measures}

\item{nBag}{a neuron bag}

\item{clusterOn}{whether to cluster on the input vector, the output vector, or both}
}
\value{
an object of class \code{connectivityCluster}. The object is a list with components:
\describe{
  \item{inputsTable}{The table of input connections to the neurons being clustered (can be NULL if clustered on outputs)}
  \item{outputsTable}{The table of output connections to the neurons being clustered (can be NULL if clustered on inputs)}
  \item{distance}{object containing a dissimilarity measure between the neurons being considered}
  \item{hc}{\code{\link{hclust}} object, result of clustering}
  \item{unit}{Unit on which the clustering was made (neuron or type, or various supertypes)}
}
}
\description{
Clustering neurons based on connectivity tables
}
\details{
\itemize{
 \item If both \code{inputsTable} and \code{outputsTable} are present, the clustering will be run on the combined connectivity vector. This is
 controlled by the \code{clusterOn} argument in \code{clusterBag}.
 \item \code{weightRelative} is used for working on input vectors, and \code{outputContribution} to work on output vectors
 }
}
\section{Functions}{
\itemize{
\item \code{clusterBag}: Cluster neurons in a bag
}}

\seealso{
\code{\link{clusterBag}}
}
