% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivityMatrices.R
\name{connectivityMatrix}
\alias{connectivityMatrix}
\title{Builds a connectivity matrix from a connection object}
\usage{
connectivityMatrix(
  connObj,
  slctROIs = NULL,
  allToAll = FALSE,
  from = "name.from",
  to = "name.to",
  value = "weightRelative",
  ref = c("inputs", "outputs")
)
}
\arguments{
\item{connObj}{a connection table}

\item{slctROIs}{which ROIs to consider. If NULL (the default), all the ROIs present in the table are used (columns of the matrix are then appended with .roi for each ROI)}

\item{allToAll}{whether to build a square matrix or just a from -> to matrix}

\item{from}{which field to use as a "source" (default "name.from")}

\item{to}{which field to use as a "target" (default "name.to")}

\item{value}{which field to use to fill the matrix (default "weightRelative")}

\item{ref}{which channel will be used as the "reference" (to be the columns of the output). The
other channel gets .roi affixed to their names in case several ROIs are considered}
}
\description{
Builds a connectivity matrix from a connection object
}
