% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivityMatrices.R
\name{cos_dist}
\alias{cos_dist}
\alias{sqrt_cos_dist}
\alias{cor_dist}
\alias{bin_dist}
\title{Distance measurements}
\usage{
cos_dist(mat)

sqrt_cos_dist(mat)

cor_dist(mat)

bin_dist(mat, threshold = 0.01)
}
\arguments{
\item{mat}{A matrix}

\item{threshold}{Threshold to use to binarize the matrix}
}
\value{
A distance object containing distances between the
rows of \code{mat}
}
\description{
Distance measurements
}
\details{
\code{cos_dist} returns the cosine distance, \code{sqrt_cos_dist}
the squared cosine distance, \code{cor_dist} one minus the spearman correlation 
between vectors, and \code{bin_dist} the binary distance after thresholding
}
\section{Functions}{
\itemize{
\item \code{sqrt_cos_dist}: Squared cosine distance

\item \code{cor_dist}: Correlation distance matrix

\item \code{bin_dist}: Binary distance
}}

