% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxtools.R
\name{cxRetyping}
\alias{cxRetyping}
\title{Convenience function to do all the CX neurons renaming}
\usage{
cxRetyping(
  connections,
  redefinePartners = TRUE,
  postfix = c("raw", "to", "from"),
  verbose = TRUE
)
}
\arguments{
\item{connections}{Connectivity table  or neuronBag to modify}

\item{redefinePartners}{If table is a neuronBag, should the partners also be retyped?}

\item{postfix}{One of "raw, "to" or "from". Specify if type (and name) columns in table to be modified are postfixed with
to and from or nothing}

\item{verbose}{Should messages about progress be displayed}
}
\description{
Convenience function to do all the CX neurons renaming
}
\details{
PFL1,3 and PFR_a and _b are renamed so that the outputs of a subtype go to the same region. All others are renamed L/R
}
