% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuronBag.R
\name{filter.neuronBag}
\alias{filter.neuronBag}
\title{Filter a neuronBag based on its \code{names} field}
\usage{
\method{filter}{neuronBag}(.nbag, filterPartners = FALSE, ...)
}
\arguments{
\item{filterPartners}{: Whether to apply the filter to input/output neurons to}

\item{...}{to be passed to a filtering function applied to the \code{names} field}

\item{nbag}{a neuronBag object to filter}
}
\description{
Filter a neuronBag based on its \code{names} field
}
