% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectionTables.R
\name{getConnectionTable}
\alias{getConnectionTable}
\title{Get connection table of inputs and add meta}
\usage{
getConnectionTable(
  bodyIDs,
  synapseType,
  slctROI = NULL,
  by.roi = FALSE,
  synThresh = 3,
  largeOnly = FALSE,
  chunk_connections = TRUE,
  chunk_meta = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{bodyIDs}{The bodyids of neurons who's connections should be queried or a metadata data.frame}

\item{synapseType}{Choose "PRE" or "POST" to get inputs or outputs of the neurons in bodyIDs, respectivly. "PRE" is
usually slower as it requires listing all the outputs of the input neurons to get the contribution of the outputs.}

\item{slctROI}{String specifying the ROI where connections should be queried. By default all the ROIs.}

\item{by.roi}{Passed to neuprint_connection_table. If returning all ROIs, should results be broken down by ROI?}

\item{synThresh}{Minimum number of synapses to consider a connection (default 3)}

\item{largeOnly}{For cropped targets, only consider them if they have more than 1000 connections (matching Scheffer et al)}

\item{chunk_meta}{: to be passed to metadata and roiInfo queries (default TRUE)}

\item{verbose}{: should the function report on its progress?}

\item{...}{Other arguments to be passed to neuprint queries (neuprint_connection_table, getRoiInfo and neuprint_get_meta)}

\item{chunk_connection}{: to be passed to neuprint_connection_table (default TRUE)}
}
\value{
Returns a connection table as data frame. Added columns are \code{weightRelativeTotal} which is
the relative weight considering all the synapses (irrespective of the ROI), and if ROI are used (either if
\code{slctROI} has a value or \code{by.roi} is \code{TRUE}), \code{weightRelative} is the relative weight in the
ROI and \code{totalROIweight} is the absolute number of inputs this neuron receives in that region and
\code{weightROIRelativeTotal} is the weight in the ROI normalized by the total number of inputs (in all ROIs)
}
\description{
Get connection table of inputs and add meta
}
