% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuronBag.R
\name{getIntraBag}
\alias{getIntraBag}
\title{Convenience to filter a neuronBag to just connections between the central neurons of a bag}
\usage{
getIntraBag(nBag)
}
\description{
Convenience to filter a neuronBag to just connections between the central neurons of a bag
}
