% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rois.R
\name{getNeuronsInRoiTable}
\alias{getNeuronsInRoiTable}
\title{Returns a table of all instances of neurons of types with non zero pre/post counts in slctROI.}
\usage{
getNeuronsInRoiTable(slctROI, minTypePercentage = 0.5)
}
\arguments{
\item{slctROI}{: the ROI to look for}

\item{minTypePercentage}{: the minimum proportion of the instances of a type that should be innervating the ROI for
it to be considered}
}
\value{
: a table of metadata for all neurons in the ROI, with extra columns \code{ROI_pre}
 and \code{ROI_post}, the counts in the queried ROI.
}
\description{
Returns a table of all instances of neurons of types with non zero pre/post counts in slctROI.
}
