% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getRoiInfo}
\alias{getRoiInfo}
\title{Formats the results of a roiInfo query in a nice table with roi/upstream/downstreams columns}
\usage{
getRoiInfo(bodyids, ...)
}
\arguments{
\item{bodyids}{A collection of bodyids}

\item{...}{To be passed to \code{neuprint_get_roiInfo}}
}
\description{
Formats the results of a roiInfo query in a nice table with roi/upstream/downstreams columns
}
