% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuronBag.R
\name{getROISummary}
\alias{getROISummary}
\title{Build a per roi summary of innervation for neurons in a neuronBag}
\usage{
getROISummary(nBag, filter = FALSE, rois = NULL)
}
\arguments{
\item{nBag}{: a neuronBag object}

\item{filter}{: if TRUE, only return results in ROIs where significant type to
type connections are found. Otherwise consider all connections (the default)}

\item{rois}{: a roiset to consider (if NULL consider all rois)}
}
\description{
Build a per roi summary of innervation for neurons in a neuronBag
}
