% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rois.R
\name{getRoiTree}
\alias{getRoiTree}
\title{Build a more useful roi table, with things ordered in a semi useful way}
\usage{
getRoiTree()
}
\description{
Build a more useful roi table, with things ordered in a semi useful way
}
