% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{getTypesTable}
\alias{getTypesTable}
\title{Returns a table with all neurons belonging to types exactly specified}
\usage{
getTypesTable(types)
}
\arguments{
\item{types:}{A vector of type names to be matched exactly}
}
\value{
A data frame of instances of those types, including a databaseType column (used internally by other functions)
}
\description{
Returns a table with all neurons belonging to types exactly specified
}
