% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{haneschPlot}
\alias{haneschPlot}
\title{Plot a hanesch type plot based on a summary of ROI innervation}
\usage{
haneschPlot(
  roiTable,
  roiSelect = selectRoiSet(getRoiTree()),
  grouping = NULL,
  flip = FALSE,
  alphaRois = 0.15,
  roiLabel = selectRoiSet(getRoiTree(), default_level = 0),
  regionOutlines = TRUE,
  theme = theme_minimal(),
  interactive = FALSE
)
}
\arguments{
\item{roiTable}{A table of ROI innervation, as generated by \code{\link{getROISummary}}}

\item{roiSelect}{A selection of ROIs table, as generated by \code{\link{selectRoiSet}}}

\item{grouping}{A string of grouping variable(s) of \code{roiTable} to facet the plot on.
Will usually be one of the \code{supertype} columns}

\item{flip}{Flip the x/y axis? If TRUE, types will be on the bottom axis and ROIs on the
side axis}

\item{alphaRois}{The alpha of the color rectangles coding the ROIs}

\item{roiLabel}{A selection of ROIs, as generated by \code{\link{selectRoiSet}} to be used for the ROI color guide}

\item{regionOutlines}{Whether or not to add the ROI color code}

\item{theme}{Theme to be passed to ggplot2}

\item{interactive}{Whether or not to use some interactivity from ggiraph}
}
\description{
Plot a hanesch type plot based on a summary of ROI innervation
}
