% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivityCluster.R
\name{is.connectivityCluster}
\alias{is.connectivityCluster}
\title{Test if x is a connectivityCluster object}
\usage{
is.connectivityCluster(x)
}
\arguments{
\item{x}{An object to be tested}
}
\value{
TRUE if x is a neuronBag
}
\description{
Test if x is a connectivityCluster object
}
