% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuronBag.R
\name{is.neuronBag}
\alias{is.neuronBag}
\title{Test if x is a neuronBag}
\usage{
is.neuronBag(x)
}
\arguments{
\item{x}{An object to be tested}
}
\value{
TRUE if x is a neuronBag
}
\description{
Test if x is a neuronBag
}
