% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retyping.R
\name{lateralize_types}
\alias{lateralize_types}
\title{Retype neurons in a table according to L/R}
\usage{
lateralize_types(
  connections,
  postfix = c("raw", "to", "from"),
  typeList = NULL,
  redefinePartners = TRUE
)
}
\arguments{
\item{connections}{Connectivity table  or neuronBag to modify}

\item{postfix}{One of "raw, "to" or "from". Specify if type (and name) columns in table to be modified are postfixed with
to and from or nothing}

\item{typeList}{: which types to lateralize (by default all the neurons which names
contains L or R)}

\item{redefinePartners}{If table is a neuronBag, should the partners also be retyped?}
}
\description{
Retype neurons in a table according to L/R
}
\examples{
\dontrun{
PFLNames <- getTypesTable(c("PFL1","PFL2","PFL3"))

## Rename only PFL2
PFLNames2 <- lateralize_types(PFLNames,postfix="raw",typeList=c("PFL2"))

##Rename all PFLs
PFLNames3 <- lateralize_types(PFLNames,postFix="raw")
}
}
