% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuprintrExtra-package.R
\docType{package}
\name{neuprintrExtra-package}
\alias{neuprintrExtra}
\alias{neuprintrExtra-package}
\title{neuprintrExtra: Extra functions to analyze neuprint EM datasets}
\description{
Analysis and display functions to complement the neuprintr package, initially developped by the Jayaraman lab to analyze
    Central complex connectivity in the hemibrain FIBSEM dataset
}
\author{
\strong{Maintainer}: Romain Franconville \email{franconviller@janelia.hhmi.org}

}
\keyword{internal}
