% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuronBag.R
\name{neuronBag}
\alias{neuronBag}
\title{neuronBag class
neuronBag objects contain information about the connectivity of a bunch of neurons}
\usage{
neuronBag(outputs, inputs, names, outputs_raw, inputs_raw, outputsTableRef)
}
\description{
neuronBag class
neuronBag objects contain information about the connectivity of a bunch of neurons
}
\details{
usually created with \code{create_neuronBag}
}
\section{Fields}{

\describe{
\item{\code{names}}{a table of metadata associated with the neurons in the bag -- with an extra column, 'databaseType'
which keeps the type name used in the database}

\item{\code{outputs}}{a type to type connectivity table of the outputs of the set of neurons}

\item{\code{inputs}}{a type to type connectivity table of the inputs of the set of neurons}

\item{\code{outputs_raw}}{a neuron to neuron connection table of the outputs of the set of neurons}

\item{\code{inputs_raw}}{a neuron to neuron connection table of the inputs of the set of neurons}

\item{\code{outputsTableRef}}{a table holding all instances of all the output types and their associated metadata}
}}

