% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typesPathways.R
\name{pathDf2graphDf}
\alias{pathDf2graphDf}
\title{Convert a type-to-type pathway dataframe to a graph dataframe}
\usage{
pathDf2graphDf(pathDf)
}
\arguments{
\item{pathDf}{Dataframe generated by get_type2typePath()}
}
\value{
Dataframe specifiying the graph resulting from combining the pathways in pathDF by listing all edges and their weights. Columns are
\code{to}, \code{from}, the statistic used as a weight in \code{pathDf} and supertypes to and from columns.
}
\description{
Convert a type-to-type pathway dataframe to a graph dataframe
}
