% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivityCluster.R
\name{plotClusters}
\alias{plotClusters}
\title{Plot a connectivityCluster object}
\usage{
plotClusters(
  connCl,
  order = NULL,
  colorAxis = FALSE,
  axesPalette = paletteer::paletteer_d("Polychrome::palette36")[3:36],
  theming = theme_minimal,
  replaceIdWithNames = TRUE,
  h = NULL,
  k = NULL
)
}
\arguments{
\item{connCl}{A \code{\link{connectivityCluster}} object}

\item{order}{Optional ordering of the units (by default the HClust order)}

\item{colorAxis}{Whether or not to color the axis labels per cluster (FALSE by default)}

\item{axesPalette}{If \code{colorAxis} is TRUE, the palette to use}

\item{theming}{A ggplot theme to use for the plot}

\item{replaceIdWithNames}{In case of neuron to neuron clustering, whether or not to replace the bodyids by neuron names on the axis labels}

\item{h}{To be passed to \code{\link{cutree}} if \code{colorAxis} is TRUE. Either \code{h} or \code{k} must be set in that case.}

\item{k}{To be passed to \code{\link{cutree}} if \code{colorAxis} is TRUE. Either \code{h} or \code{k} must be set in that case.}
}
\description{
Plot a connectivityCluster object
}
