% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotConnectivity}
\alias{plotConnectivity}
\title{Plot a connectivity matrix}
\usage{
plotConnectivity(
  connObj,
  slctROI = NULL,
  grouping = "type",
  connectionMeasure = "weightRelative",
  xaxis = c("inputs", "outputs"),
  facetInputs = NULL,
  facetOutputs = NULL,
  theme = theme_minimal(),
  cmax = NULL,
  replacementLabels = NULL,
  orderIn = NULL,
  orderOut = NULL,
  legendName = NULL,
  showTable = "inputs",
  switch = NULL
)
}
\arguments{
\item{connObj}{A connectivity object. Either a connectivity table, a matrix with \code{dimnames} Inputs and Outputs(as returned by \code{\link{connectivityMatrix}}) or a \code{\link{connectivityCluster}} object.}

\item{slctROI}{For connectivity tables, you can specify a ROI}

\item{grouping}{Which variable to use. Will be ignored when \code{connObj} is a \code{\link{connectivityCluster}} 
(the decision has already been made while clustering). Any variable postfixed by "to" or "from" in the table is a valid value, as well as "bodyid" and "neuron"}

\item{connectionMeasure}{Which variable to use as a plotting value}

\item{xaxis}{Should inputs or outputs be on the x axis (will match the value to "inputs" or "outputs")}

\item{facetInputs}{Variable to facet the inputs on (nothing by default)}

\item{facetOutputs}{Variable to facet the outputs on (nothing by default)}

\item{theme}{A theme to use}

\item{cmax}{Maximum fill value for the color scale. By default the maximum value found in the table.}

\item{replacementLabels}{A column prefix to use as a replacement for the axis labels (useful to replace bodyids with names for example)}

\item{orderIn}{A vector of input bodyid/types/names in the desired order, or a connectivityCluster object. 
Optional ordering of the inputs (ignored and replaced by the clustering order if connObj is a connectivityCluster).}

\item{orderOut}{A vector of output bodyid/types/names in the desired order, or a connectivityCluster object. 
Optional ordering of the outputs (ignored and replaced by the clustering order if connObj is a connectivityCluster).}

\item{legendName}{Optional override the default name for the color legend (by default a prettification of connectionMeasure)}

\item{showTable}{When both inputs and outputs have been used for a clustering (via \code{\link{clusterBag}}), which connectivity table to show.}

\item{switch}{To be passed to \code{\link{ggplot2::facet_grid}} to switch where the facet labels are displayed.}
}
\value{
A ggplot object
}
\description{
Plot a connectivity matrix
}
\details{
orderIn and orderOut are passed as levels to a factor to order the axis.
}
