% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectionTables.R
\name{processConnectionTable}
\alias{processConnectionTable}
\title{Internal function, exposed only for fringe cases like comparing different versions of the dataset.}
\usage{
processConnectionTable(
  myConnections,
  synThresh,
  refMeta,
  partnerMeta,
  refMetaOrig,
  synapseType,
  by.roi,
  slctROI,
  verbose,
  chunk_meta,
  chunk_connections,
  ...
)
}
\description{
Internal function, exposed only for fringe cases like comparing different versions of the dataset.
}
