% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retyping.R
\name{redefineTypeByBodyId}
\alias{redefineTypeByBodyId}
\title{Retype according to sets of bodyids}
\usage{
redefineTypeByBodyId(
  connections,
  sets,
  nameModifiers,
  postfix = c("raw", "to", "from"),
  redefinePartners = FALSE
)
}
\arguments{
\item{connections}{Connectivity table  or neuronBag to modify}

\item{sets}{List of neuron bodyid sets.}

\item{nameModifiers}{Character vector of length the same length as sets containing}

\item{postfix}{One of "raw, "to" or "from". Specify if type (and name) columns in table to be modified are postfixed with
to and from or nothing}

\item{redefinePartners}{If table is a neuronBag, should the partners also be retyped?}
}
\description{
Retype according to sets of bodyids
}
