% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retyping.R
\name{redefineTypeBySet}
\alias{redefineTypeBySet}
\title{Retype according to sets of names}
\usage{
redefineTypeBySet(
  connections,
  postfix,
  redefinePartners = TRUE,
  sets,
  nameModifiers
)
}
\arguments{
\item{connections}{Connectivity table  or neuronBag to modify}

\item{postfix}{One of "raw, "to" or "from". Specify if type (and name) columns in table to be modified are postfixed with
to and from or nothing}

\item{redefinePartners}{If table is a neuronBag, should the partners also be retyped?}

\item{sets}{List of neuron name sets}

\item{nameModifiers}{Character vector of length the same length as sets containing postfixes
to be appended to the type name of all neurons whose name is in the corresponding set}
}
\description{
Retype according to sets of names
}
