% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rois.R
\name{roiOutline}
\alias{roiOutline}
\title{Get 2D outlines for a ROI}
\usage{
roiOutline(roiMesh, alpha = 100, roiName)
}
\arguments{
\item{roiMesh}{: either a mesh3d object (as returned by \code{neuprint_ROI_mesh}) or the name of an
existing ROI}

\item{alpha}{: alpha parameter to be passed to the \code{alphahull::ahull} function}

\item{roiName}{: a name to label the outline with}
}
\value{
a data frame with columns x, y (outlines) proj (either xy or xz, the projection considered) and roi (the roi name)
}
\description{
Get 2D outlines for a ROI
}
\details{
both xy and xz projections are returned, which can be selected with the \code{proj} column. This is meant to
be handy for ggplot2 plotting
}
