% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rois.R
\name{roisPalette}
\alias{roisPalette}
\title{Create a consistent palette for ROIs for plotting}
\usage{
roisPalette(rois = getRoiTree(), favoriteRegion = "CX", my_palette)
}
\arguments{
\item{rois}{A ROI hierarchy as returned by \code{getRoiTree}}

\item{favoriteRegion}{A brain region (defined at level 1) for which one wants colors
defined down to level 2}

\item{my_palette}{A discrete color palette to use}
}
\value{
A named vector of colors (for example to be used in ggplot2's \code{scale_color_manual})
}
\description{
Create a consistent palette for ROIs for plotting
}
\seealso{
\code{getRoiTree}
}
