% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxtools.R
\name{scale_color_CX_supertype}
\alias{scale_color_CX_supertype}
\title{A color scale for Central complex supertypes}
\usage{
scale_color_CX_supertype(...)
}
\value{
A ggplot scale to be used in plots were color (or fill) map to level 2 supertypes
}
\description{
A color scale for Central complex supertypes
}
