% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supertypeUtils.R
\name{selectSupertypeSet}
\alias{selectSupertypeSet}
\title{Select a set of Supertypes from a Supertypes table as the one generated by supertype (when applied to a data frame)}
\usage{
selectSupertypeSet(
  supertypeTable,
  default_level = 2,
  exceptions = NULL,
  exceptionLevelMatch = default_level
)
}
\arguments{
\item{supertypeTable}{A supertypes table as returned by \code{supertype}}

\item{default_level}{An integer specifying which Supertype level to use by default}

\item{exceptions}{A list of the form list(nameOfType = level) where nameOfType is the name of
Supertypes at level \code{exceptionLevelMatch} for which one want to use a different level of description}

\item{exceptionLevelMatch}{What level to use in the \code{exceptions} list}
}
\value{
a data.frame with the same columns as those returned by \code{supertype} with an extra \code{type} column
containing the desired set
}
\description{
Select a set of Supertypes from a Supertypes table as the one generated by supertype (when applied to a data frame)
}
\details{
this is to supertypes what \code{selectRoiSet} is to ROIs
}
\seealso{
\code{supertype}, \code{selectRoiSet}
}
