% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivityCluster.R
\name{setClusters}
\alias{setClusters}
\alias{addClusters}
\title{Add a cluster column to the tables of a connectivityCluster object 
Add a cluster column to the tables of a connectivityCluster object, either for the neurons being clustered (with \code{setClusters}) or for the partner neurons 
(with \code{addClusters})}
\usage{
setClusters(connCl, h = 0.8, k = NULL)

addClusters(connCl, partnerClusters, h = 0.8, k = NULL)
}
\arguments{
\item{connCl}{A connectivityCluster object to augment}

\item{h}{To be passed to \code{\link{cutree}} to define the clusters}

\item{k}{To be passed to \code{\link{cutree}} to define the clusters}

\item{partnerClusters}{Either a named vector (as returned by cutree) or a \code{\link{connectivityCluster}} object defining a clustering on the partner neurons in
\code{connCl}}
}
\value{
A connectivityCluster, whith added columns cluster.from and cluster.to in the corresponding connectivity table
}
\description{
Add a cluster column to the tables of a connectivityCluster object 
Add a cluster column to the tables of a connectivityCluster object, either for the neurons being clustered (with \code{setClusters}) or for the partner neurons 
(with \code{addClusters})
}
\section{Functions}{
\itemize{
\item \code{addClusters}: Add clustering of the partners
}}

