% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectionTables.R
\name{simplifyConnectionTable}
\alias{simplifyConnectionTable}
\title{Change from a bodyid/partner/prepost to a from/to format}
\usage{
simplifyConnectionTable(connectionTable)
}
\arguments{
\item{connectionTable:}{A data frame in the bodyid/partner/prepost format (as returned by getConnectionTable)}
}
\value{
A data frame in the from/to/name.from/name.to... format
}
\description{
Change from a bodyid/partner/prepost to a from/to format
}
