% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supertypeUtils.R
\name{supertype.character}
\alias{supertype.character}
\title{Gets a supertype from a type name, with various levels of depth}
\usage{
\method{supertype}{character}(types, level = 2)
}
\arguments{
\item{types}{: a type name or vector of type names}

\item{level}{: depth of the supertype. Possible values are 1,2 or 3 (default 2), 1 being the finest 
subdivision and 3 the coarsest}
}
\description{
Gets a supertype from a type name, with various levels of depth
}
\details{
For example, at level 1 delta0 neurons are just divided in DeltaA to K, at level 2 they are
D0, and at level 3 they are FB Interneurons
}
