% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cxtools.R
\name{supertype2Palette}
\alias{supertype2Palette}
\title{A color scale for Central complex supertypes}
\usage{
supertype2Palette()
}
\value{
A list of colors (and breaks) mapping CX supertypes to palette36 colors
}
\description{
A color scale for Central complex supertypes
}
