% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typesPathways.R
\name{tableChain2path}
\alias{tableChain2path}
\title{Link an arbitrary number of connection tables into a pathway table}
\usage{
tableChain2path(
  ...,
  n_steps = NULL,
  stat = NULL,
  excludeLoops = TRUE,
  type.to = NULL
)
}
\arguments{
\item{...}{An arbitrary number of connection tables or a list of connection tables. In desired downstream order}

\item{stat}{which stat to use for pathway weights. Can be a vector. If NULL (the default), all the available meaningful stats are used: 
(\code{"weightRelative","outputContribution","knownWeightRelative","knownOutputContribution","knownOutputContribution_perType","knownWROutputContribution_perType","knownWeightRelative_perType"})}

\item{excludeLoops}{Logical, whether to exclude paths containing duplicates (TRUE by default)}

\item{type.to}{a metadata data.frame for a set of neurons (typically generated by a search or a getTypesTable call) to arrive to}
}
\description{
Link an arbitrary number of connection tables into a pathway table
}
