% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rois.R
\name{typesInROI}
\alias{typesInROI}
\title{Extract types participating in a significant connection in a ROI}
\usage{
typesInROI(roiConnections, ROI)
}
\arguments{
\item{roiConnections}{A \code{neuronBag} object}

\item{ROI}{The ROI to consider}
}
\value{
A dataframe with columns "type" and "databaseType"
}
\description{
Extract types participating in a significant connection in a ROI
}
