% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supertypeUtils.R
\name{typesPalette}
\alias{typesPalette}
\title{Create a consistent palette for supertypes for plotting}
\usage{
typesPalette(typeSelection, my_palette = paletteer_d("Polychrome::palette36"))
}
\arguments{
\item{typeSelection}{A supertype selection as returned by \code{selectSupertypeSet}}

\item{my_palette}{A discrete color palette to use}
}
\value{
A named vector of colors (for example to be used in ggplot2's \code{scale_color_manual})
}
\description{
Create a consistent palette for supertypes for plotting
}
\seealso{
\code{selectSupertypeSet}, \code{roisPalette}
}
