% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retyping.R
\name{get_col_name}
\alias{get_col_name}
\title{Small utility to generate "type.from" kind of names}
\usage{
get_col_name(
  col = c("bodyid", "name", "type", "databaseType", "supertype", paste0("supertype",
    1:3)),
  post = c("raw", "from", "to")
)
}
\description{
Small utility to generate "type.from" kind of names
}
