% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typesPathways.R
\name{get_type2typePath}
\alias{get_type2typePath}
\title{Builds a data frame of pathways between 2 collections of types}
\usage{
get_type2typePath(
  type.from,
  type.to,
  by.roi = TRUE,
  ROI = NULL,
  n_steps = 2,
  renaming = cxRetyping,
  stat = "weightRelative",
  ...
)
}
\arguments{
\item{type.from}{a metadata data.frame for a set of neurons (typically generated by a search or a getTypesTable call) to start from}

\item{type.to}{a metadata data.frame for a set of neurons (typically generated by a search or a getTypesTable call) to arrive to}

\item{by.roi}{return results by ROI or just global weights?}

\item{ROI}{which ROIs to consider. Can be a vector of ROI names, a list of vectors of ROI names, or NULL. If NULL (the default), consider all ROIs. 
If a character vector, restrict to those ROIs. If a list, each element of the list will be ROI that will be combined together (and named as the element name in the list) using combineRois}

\item{...}{: to be passed to create_neuronBag when building the path}
}
\description{
Builds a data frame of pathways between 2 collections of types
}
