% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neuronBag.R
\name{getROISummary}
\alias{getROISummary}
\title{Build a per roi summary of innervation for neurons in a neuronBag}
\usage{
getROISummary(neurons, threshold = 0, rois = NULL)
}
\arguments{
\item{neurons}{: a dataframe as returned by a search or a neuronBag object}

\item{threshold}{: the minimal average number of synapses for a type in a ROI
for it to be included}

\item{rois}{: a roiset to consider (if NULL consider all rois)}
}
\description{
Build a per roi summary of innervation for neurons in a neuronBag
}
