% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rois.R
\name{getRoiTree}
\alias{getRoiTree}
\title{A wrapper around \code{neuprint_ROI_hierarchy} returning a cleaned up, sorted table with various ROI levels.}
\usage{
getRoiTree()
}
\value{
A table with columns `level` 0 through 4, 0 being the more general and 4 the more detailed, omitting tracts. ROI are somewhat sorted,
 from right to central to left (and from periphery to center to periphery). level0 is similar to level1 except that Left/Right distinctions are omitted.
}
\description{
A wrapper around \code{neuprint_ROI_hierarchy} returning a cleaned up, sorted table with various ROI levels.
}
\seealso{
\code{selectRoiSet} to make a selection of ROIs (for example for a figure) from such a hierarchy
}
