% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rois.R
\name{getTypesInRoiTable}
\alias{getTypesInRoiTable}
\title{Returns a neuronBag object of all the neurons forming significant connections in a ROI.}
\usage{
getTypesInRoiTable(
  ROI,
  renaming = cxRetyping,
  bagROIs = ROI,
  minTypePercentage = 0.5,
  ...
)
}
\arguments{
\item{ROI}{The ROI to consider}

\item{renaming}{a renaming function to be applied to the types found. \code{cxRetyping} by default (for no renaming, set it to \code{identity})}

\item{bagROIs}{Which ROIs to include in the bag created (by default only the ROI one wants neurons in). If NULL returns all ROIs.}

\item{minTypePercentage}{The minimum proportion of the instances of a type that should be innervating the ROI for
it to be considered (0.5 by default)}

\item{...}{Parameters to be passed to \code{neuronBag}}
}
\description{
Returns a neuronBag object of all the neurons forming significant connections in a ROI.
}
\details{
calls \code{getNeuronsInRoiTable} internally, with \code{minTypePercentage} divided by 2 if \code{lateralize} is TRUE.
}
\seealso{
\code{getNeuronsInRoiTable}
}
