% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rois.R
\name{getTypesInRoiTable}
\alias{getTypesInRoiTable}
\title{Returns a neuronBag object of all the neurons in the ROI}
\usage{
getTypesInRoiTable(ROI, lateralize = FALSE, ...)
}
\arguments{
\item{ROI}{: the ROI to consider}

\item{lateralize}{: should the neuron types be divided in left/right (default FALSE)}

\item{big}{: if TRUE, run through a pblapply call}

\item{clN}{: if big is TRUE, this is the number of cores to use.}
}
\description{
Returns a neuronBag object of all the neurons in the ROI
}
