% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rois.R
\name{getTypesInRoiTable}
\alias{getTypesInRoiTable}
\title{Returns a neuronBag object of all the neurons forming significant connections in a ROI}
\usage{
getTypesInRoiTable(ROI, lateralize = FALSE, ...)
}
\arguments{
\item{ROI}{: the ROI to consider}

\item{lateralize}{: should the neuron types be divided in left/right (default FALSE)}

\item{...}{: parameters to be passed to}
}
\description{
Returns a neuronBag object of all the neurons forming significant connections in a ROI
}
\details{
calls \code{getNeuronsInRoiTable} internally, with \code{minTypePercentage} set to 0.5, or 0.25 if \code{lateralize} is TRUE.
}
\seealso{
\code{getNeuronsInRoiTable}
}
