% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connectivityCluster.R
\name{plotClusters}
\alias{plotClusters}
\title{Plot a connectivityCluster object}
\usage{
plotClusters(
  connCl,
  orderX = connCl$hc$order,
  orderY = connCl$hc$order,
  colorAxis = FALSE,
  axesPalette = paletteer::paletteer_d("Polychrome::palette36")[3:36],
  theme = theme_minimal(),
  replacementLabels = NULL,
  h = NULL,
  k = NULL
)
}
\arguments{
\item{connCl}{A \code{\link{connectivityCluster}} object}

\item{orderX}{Optional ordering of the groupings (by default the HClust order)}

\item{orderY}{Optional ordering of the groupings (by default the HClust order)}

\item{colorAxis}{Whether or not to color the axis labels per cluster (FALSE by default)}

\item{axesPalette}{If \code{colorAxis} is TRUE, the palette to use}

\item{theme}{A ggplot theme to use for the plot}

\item{replacementLabels}{A column prefix to use as a replacement for the axis labels (useful to replace bodyids with names for example)}

\item{h}{To be passed to \code{\link{cutree}} if \code{colorAxis} is TRUE. Either \code{h} or \code{k} must be set in that case.}

\item{k}{To be passed to \code{\link{cutree}} if \code{colorAxis} is TRUE. Either \code{h} or \code{k} must be set in that case.}
}
\description{
Plot a connectivityCluster object
}
