% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plotConnectivity}
\alias{plotConnectivity}
\title{Plot a connectivity matrix}
\usage{
plotConnectivity(
  connObj,
  slctROIs = NULL,
  grouping = c("type", "neuron", "supertype1", "supertype2", "supertype3",
    "databaseType"),
  value = "weightRelative",
  xaxis = c("inputs", "outputs"),
  theme = theme_minimal(),
  cmax = NULL,
  replaceIds = TRUE,
  orderIn = NULL,
  orderOut = NULL,
  showTable = "inputs"
)
}
\arguments{
\item{connObj}{A connectivity object. Either a connectivity table, a matrix with \code{dimnames} Inputs and Outputs(as returned by \code{\link{connectivityMatrix}}) or a \code{\link{connectivityCluster}} object.}

\item{slctROIs}{For connectivity tables, you can specify a selection of ROIs}

\item{grouping}{Which variable to use. Will be ignored when \code{connObj} is a \code{\link{connectivityCluster}} (the decision has already been made while clustering)}

\item{value}{Which variable to use as a plotting value}

\item{xaxis}{Should inputs or outputs be on the x axis (will match the value to "inputs" or "outputs")}

\item{theme}{A theme to use}

\item{cmax}{Maximum fill value for the color scale}

\item{replaceIds}{When plotting neuron to neuron connections: if TRUE (and connObj is either a table or a connectivityCluster), replace the bodyids with their corresponding name. For matrices also accepts a vector of names}

\item{orderIn}{Optional ordering of the inputs (ignored and replaced by the clustering order if connObj is a connectivityCluster).}

\item{orderOut}{Optional ordering of the outputs (ignored and replaced by the clustering order if connObj is a connectivityCluster).}

\item{showTable}{When both inputs and outputs have been used for a clustering, which connectivity table to show.}
}
\description{
Plot a connectivity matrix
}
