% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retyping.R
\name{redefine_types}
\alias{redefine_types}
\title{Redefine types in a table/neuronBag}
\usage{
redefine_types(
  connections,
  retype_func,
  postfix = c("raw", "from", "to"),
  redefinePartners = TRUE,
  ...
)
}
\arguments{
\item{connections}{A connection table or a neuronBag object}

\item{retype_func}{A function taking \code{connections} and \code{postfix} as argument
generating new type names from the table}

\item{postfix}{One of "raw, "to" or "from". Specify if type (and name) columns in table to be modified are postfixed with
to and from or nothing}

\item{redefinePartners}{If table is a neuronBag, should the partners also be retyped?}

\item{...}{Extra parameters to be passed to retype_func}
}
\value{
A data frame or neuronBag with the columns \code{type_col} updated.
}
\description{
Redefine types in a table/neuronBag
}
\details{
This is a low level functions. In most use cases, you can use \code{lateralize_types} or \code{redefineTypeByName} instead
}
\seealso{
\code{\link{lateralize_types}}, \code{\link{redefineTypeByName}}
}
