% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retyping.R
\name{retype.na}
\alias{retype.na}
\title{Fill in the type and name field in case they are NAs, using the name field if it exists
(removing the _L/_R) or the neuron id. By default expects a table in to/from format.}
\usage{
retype.na(connectionTable)
}
\description{
Fill in the type and name field in case they are NAs, using the name field if it exists
(removing the _L/_R) or the neuron id. By default expects a table in to/from format.
}
