% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rois.R
\name{selectRoiSet}
\alias{selectRoiSet}
\title{Select a set of ROIs from a ROI hierarchy as the one generated by getRoiTree}
\usage{
selectRoiSet(
  roiTree = getRoiTree(),
  default_level = 2,
  exceptions = NULL,
  exceptionLevelMatch = default_level
)
}
\arguments{
\item{roiTree}{A ROI hierarchy table as returned by \code{getRoiTree}}

\item{default_level}{An integer specifying which ROI level to use by default}

\item{exceptions}{A list of the form list(nameOfRoi = level) where nameOfRoi is the name of
ROIs at level \code{exceptionLevelMatch} for which one want to use a different level of description}

\item{exceptionLevelMatch}{What level to use in the \code{exceptions} list. Can be either a scalar or a vector the same
length as \code{exceptions}}
}
\value{
a data.frame with the same columns as those returned by \code{getRoiTree} with an extra \code{roi} column
containing the desired set
}
\description{
Select a set of ROIs from a ROI hierarchy as the one generated by getRoiTree
}
\seealso{
\code{getRoiTree}
}
