% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typesPathways.R
\name{tableChain2path}
\alias{tableChain2path}
\title{Link an arbitrary number of connection tables into a pathway table}
\usage{
tableChain2path(
  ...,
  n_steps = NULL,
  stat = "weightRelative",
  excludeLoops = TRUE
)
}
\arguments{
\item{...}{An arbitrary number of connection tables or a list of connection tables. In desired downstream order}

\item{stat}{the stat to consider (by default "weightRelative")}

\item{excludeLoops}{Logical, whether to exclude paths containing duplicates (TRUE by default)}
}
\description{
Link an arbitrary number of connection tables into a pathway table
}
