% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/typesPathways.R
\name{tables2path}
\alias{tables2path}
\title{Multiply 2 connection tables or a pathway table and a connection table and return a type to type pathway table}
\usage{
tables2path(inputTable, outputTable, stat = "weightRelative", n = 1)
}
\arguments{
\item{inputTable}{a type to type connection table, or a pathway table as returned by this function}

\item{outputTable}{a type to type connection table}

\item{stat}{which stat to use for pathway weights. Can be a vector.}

\item{n}{a number to label the stage in processing}
}
\value{
a table with a stat_pathway extra column, and type and roi columns labeled with n or n+1 depending if they are 
at the first or second step of the pathway
}
\description{
Multiply 2 connection tables or a pathway table and a connection table and return a type to type pathway table
}
